Li-ion Battery Dataset (NASA)

These data are from four batteries, which were instrumented to capture the value of variables such as current, temperature and voltage from sensors. The batteries were subjected to charging and discharging processes in a controlled environment with a constant ambient temperature. The batteries are named like this:

Battery 5

Battery 6

Battery 7

Battery 18

data_NASA

Format

data_NASA

A data frame with 185.721 rows and 12 columns:

cycle

Battery discharge cycle number

ambient_temperature

Ambient temperature (degree C)

datetime

The date and time of the start of the cycle

capacity

Battery capacity (Ahr)

voltage_measured

Battery terminal voltage (Volts)

current_measured

Battery output current (Amps)

temperature_measured

Battery temperature (degree C)

current_load

Current measured at load (Amps)

voltage_load

Voltage measured at load (Volts)

time

Time vector for the cycle (secs)

soh

Percentage of battery health status in each cycle

battery

Battery identification number

Source

https://data.nasa.gov/